webapp - refactor: SectorPieChart 메인 컴포넌트에 있던 유틸 함수를 하위의 utils 디렉토리로 옮김
