+++ packages/webapp/src/pages/Dashboard/SectorPieChart/utils/calcSectorRatios.ts
@@ -1,0 +1,14 @@
+import { SectorPieChartRatio } from '@types';
+
+export default function calcSectorRatios(
+	sectorMap: Map<string, string[]>,
+	numOfHoldings: number
+): SectorPieChartRatio[] {
+	return [...sectorMap]
+		.map(([sector, tickers]) => ({
+			sector,
+			ratio: tickers.length / numOfHoldings,
+			includedStocks: [...tickers]
+		}))
+		.sort((a, b) => b.ratio - a.ratio);
+}
