+++ packages/webapp/src/pages/Dashboard/SectorPieChart/utils/index.ts
@@ -1,2 +1,5 @@
+export { default as calcSectorRatios } from './calcSectorRatios';
+export { default as convertToSectorChartData } from './convertToSectorChartData';
 export { default as drawPieChart } from './drawPieChart';
+export { default as initializeSectorMap } from './initializeSectorMap';
 export { default as translateSectorToKor } from './translateSectorToKor';
